+++ Week11/VerifyDate.h
@@ -9,16 +9,16 @@ private:
 	// Data Members
 	int month = 01;
 	int day = 01;
-	double year = 2000;
+	int year = 2000;
 
 	// Private Mutator Functions
-	bool isLeap(double) const;
+	bool isLeap(int) const;
 	
 public:
 	// Setters
 	void setDay(int);
 	void setMonth(int);
-	void setYear(double);
+	void setYear(int);
 
 	// Getters
 	int getDay() const;
@@ -27,8 +27,8 @@ public:
 
 	// Public Mutator Functions
 	string printDate();
-	bool validateDate(int, int, double);
+	bool validateDate(int, int, int);
 	bool validateMonth(int);
-	bool validateYear(double);
+	bool validateYear(int);
 };
